fixed lineage
