+++ frontend/src/components/ui/WorkspaceFetcher.jsx
@@ -11,71 +11,6 @@ import { activeConfigurationAtom } from "@/atoms/anvilConfigurationsAtom";
 import { fetchExecutionDetails, getAllPipelines, ping } from "@/client/anvil";
 import { useSyncExecutionResults } from "@/hooks/useExecutionResults";
 
-const getLineage = (workspace) => {
-  const lineage = new Map();
-
-  if (!workspace?.pipelines) {
-    return lineage;
-  }
-
-  // Filter out pipelines with empty .record fields
-  const validPipelines = Object.values(workspace?.pipelines).filter(
-    (pipeline) => pipeline.record && Object.keys(pipeline.record).length > 0,
-  );
-
-  Object.values(validPipelines).forEach((pipeline) => {
-    const record = pipeline?.record;
-    const sha1Hash = record?.Hash;
-    if (!record || !sha1Hash) {
-      return;
-    }
-    const pipelineData = JSON.parse(record.PipelineJson);
-    const friendlyName = pipelineData.name;
-    if (!lineage.has(sha1Hash)) {
-      const createDate = new Date(record.Created * 1000);
-      lineage.set(sha1Hash, {
-        id: pipeline.id,
-        name: friendlyName,
-        hash: sha1Hash,
-        deployed: record.Deployed,
-        pipelineData: pipelineData,
-        created: createDate.toLocaleString(),
-        lastExecution: createDate.toLocaleString(),
-        host: pipeline.host,
-        executions: new Map(),
-      });
-    }
-
-    const lineageEntry = lineage.get(sha1Hash);
-    const existingExecution = lineageEntry.executions.get(record.Execution);
-
-    if (!existingExecution) {
-      const execDate = new Date(record.ExecutionTime * 1000);
-      lineageEntry.executions.set(record.Execution, {
-        id: record.Execution,
-        hash: sha1Hash,
-        pipeline: pipeline.id,
-        created: execDate.toLocaleString(),
-        status: record.Status,
-      });
-    } else {
-      Object.assign(existingExecution, {
-        status: pipeline.record.Status,
-        created: new Date(
-          pipeline.record.ExecutionTime * 1000,
-        ).toLocaleString(),
-      });
-    }
-
-    const mostRecent = Array.from(lineageEntry.executions.values())[0]?.created;
-    if (mostRecent) {
-      lineageEntry.lastExecution = mostRecent;
-    }
-  });
-
-  return lineage;
-};
-
 export default function WorkspaceFetcher() {
   const [workspace, setWorkspace] = useImmerAtom(workspaceAtom);
   const loadPipeline = useLoadServerPipeline();
@@ -157,31 +92,36 @@ export default function WorkspaceFetcher() {
       }),
   });
 
-  const loadShell = async (key, serverPipeline, configuration) => {
-    const loaded = await loadPipeline(serverPipeline, configuration);
-    setWorkspace((draft) => {
-      draft.pipelines[key] = loaded;
-    });
-  };
-
   useEffect(() => {
-    if (pipelinesData) {
-      pipelinesData.body?.forEach((serverPipeline) => {
-        const key = serverPipeline.Uuid + "." + serverPipeline.Execution;
-        const existing = workspace.pipelines[key];
-        if (!existing) {
-          loadShell(key, serverPipeline, pipelinesData.configuration);
-        }
-      });
+    if (!pipelinesData?.body) {
+      return;
     }
-  }, [pipelinesData]);
-
-  useEffect(() => {
-    const lineage = getLineage(workspace);
-    setWorkspace((draft) => {
-      draft.lineage = lineage;
+    const pipelinesToLoad = pipelinesData.body?.filter((serverPipeline) => {
+      const key = serverPipeline.Uuid + "." + serverPipeline.Execution;
+      return !workspace.pipelines[key];
     });
-  }, [workspace.pipelines]);
+
+    // If there are pipelines to load, process them
+    const loadPromises = pipelinesToLoad.map(async (serverPipeline) => {
+      const key = serverPipeline.Uuid + "." + serverPipeline.Execution;
+      const loaded = await loadPipeline(
+        serverPipeline,
+        pipelinesData.configuration,
+      );
+      return [key, loaded];
+    });
+
+    Promise.all(loadPromises)
+      .then((results) => {
+        const loadObj = Object.fromEntries(results);
+        setWorkspace((draft) => {
+          draft.pipelines = { ...workspace.pipelines, ...loadObj };
+        });
+      })
+      .catch((error) => {
+        console.error("Error loading pipelines:", error);
+      });
+  }, [pipelinesData]);
 
   return null;
 }
